+++ src/sbt-test/rpm/config-no-replace-dsl/src/universal/conf/test
@@ -1,0 +1,1 @@
+# Test configuration file!
