+++ src/sbt-test/rpm/config-no-replace-dsl/test
@@ -1,0 +1,7 @@
+# Run the debian packaging.
+> rpm:package-bin
+$ exists target/rpm/RPMS/x86_64/rpm-test-0.1.0-1.x86_64.rpm
+
+> unzip
+
+> checkSpecFile
